requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.2
Django==5.2.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
isort==6.0.1
kombu==5.5.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
